+++ custom_components/hikvision_intercom/const.py
@@ -1,7 +1,7 @@
 """Shared integration constants."""
 
 DOMAIN = "hikvision_intercom"
-VERSION = "1.0.0-rc.4"
+VERSION = "1.0.0-rc.5"
 PLATFORMS = ("binary_sensor", "sensor", "camera", "lock", "event")
 DEFAULT_IDLE_INTERVAL = 2.0
 DEFAULT_ACTIVE_INTERVAL = 0.75